culsh/python/requirements.txt
numpy~=1.0
cupy-cuda12x
